下準備リファクタ
